Enrico Backend registro, fornecedor e fatura

+++ EchoVision_spring/EchoVision/src/main/java/br/com/EchoVision/EchoVision/repository/FaturaRepository.java
@@ -1,9 +1,10 @@
-package br.com.EchoVision.EchoVision.repository; 
+package br.com.EchoVision.EchoVision.repository;
 
 import br.com.EchoVision.EchoVision.models.Fatura;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FaturaRepository extends JpaRepository<Fatura, Long> {
+public interface FaturaRepository extends JpaRepository<Fatura, Integer> {
+    // O Spring já gera os métodos de salvar e buscar compras automaticamente aqui
 }
