+++ EchoVision_spring/EchoVision/src/main/java/br/com/EchoVision/EchoVision/repository/RegistroRepository.java
@@ -1,9 +1,9 @@
-package br.com.EchoVision.EchoVision.repository; 
+package br.com.EchoVision.EchoVision.repository;
 
 import br.com.EchoVision.EchoVision.models.Registro;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegistroRepository extends JpaRepository<Registro, Long> {
+public interface RegistroRepository extends JpaRepository<Registro, Integer> {
 }
